refactor(core): search.rs のテストを機能別モジュールへ分割 (#597)
search.rs の `#[cfg(test)] mod tests`（約 1,615 行）を責務単位の
`search/tests/` サブモジュールへ物理分割する。製品コードは一切変更せず、
テスト本体はスクリプト抽出でバイト等価に移設した。

- 索引 mod.rs + 共通 fixture common.rs + 機能別 6 ファイル
  （basic / ranking / incremental / migemo / path / performance）
- テスト名・ignored 属性は不変（葉のテスト名 80 件が分割前と完全一致・
  74 通常 / 6 ignored）。`bench_` フィルタも従来どおり 4 件を列挙
- サブモジュールは `crate::search` の descendant として private 実装
  （フィールド・ScoredEntry・adjusted_history_boost 等）へ直接アクセス
- snotra-core/CLAUDE.md の search.rs 索引に test サブモジュール群を追記
  （責務散文は各ファイルの //! を正準・#562）

検証: 葉テスト名の分割前後 diff 一致 / cargo test -p snotra-core（74 passed,
6 ignored）/ clippy / rustdoc(-D warnings) / governance:check G1..G9。
