+++ snotra-core/src/search.rs
@@ -994,1617 +994,4 @@ fn match_score_single_cached(
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::history::HistoryStore;
-    use crate::indexer::AppEntry;
-
-    fn make_entries(names: &[&str]) -> Vec<AppEntry> {
-        names
-            .iter()
-            .map(|n| AppEntry {
-                name: n.to_string(),
-                target_path: format!("C:\\fake\\{}.lnk", n),
-                is_folder: false,
-            })
-            .collect()
-    }
-
-    fn empty_history() -> HistoryStore {
-        HistoryStore::load()
-    }
-
-    #[test]
-    fn search_empty_query_returns_empty() {
-        let mut engine = SearchEngine::new(make_entries(&["Firefox", "Chrome"]));
-        let results = engine.search("", 8, &empty_history(), SearchMode::Fuzzy);
-        assert!(results.is_empty());
-    }
-
-    #[test]
-    fn search_no_entries_returns_empty() {
-        let mut engine = SearchEngine::new(Vec::new());
-        let results = engine.search("fire", 8, &empty_history(), SearchMode::Fuzzy);
-        assert!(results.is_empty());
-    }
-
-    #[test]
-    fn search_returns_fuzzy_matches() {
-        let entries = make_entries(&["Firefox", "Chrome", "Notepad", "Visual Studio Code"]);
-        let mut engine = SearchEngine::new(entries);
-        let results = engine.search("fire", 8, &empty_history(), SearchMode::Fuzzy);
-        assert!(!results.is_empty());
-        assert_eq!(results[0].name, "Firefox");
-    }
-
-    #[test]
-    fn search_respects_max_results() {
-        let entries = make_entries(&["app1", "app2", "app3", "app4", "app5"]);
-        let mut engine = SearchEngine::new(entries);
-        let results = engine.search("app", 3, &empty_history(), SearchMode::Fuzzy);
-        assert!(results.len() <= 3);
-    }
-
-    #[test]
-    fn search_top_k_replaces_worst_when_better_arrives_late() {
-        let entries = make_entries(&[
-            "appabcdefghij",
-            "appabcdefghi",
-            "appx",
-            "app",
-        ]);
-        let mut engine = SearchEngine::new(entries);
-
-        let results = engine.search("app", 2, &empty_history(), SearchMode::Prefix);
-        let names: Vec<&str> = results.iter().map(|r| r.name.as_str()).collect();
-
-        assert_eq!(names, vec!["app", "appx"]);
-    }
-
-    #[test]
-    fn search_top_k_is_independent_from_input_order() {
-        let ordered = make_entries(&[
-            "appabcdefghij",
-            "appabcdefghi",
-            "appx",
-            "app",
-        ]);
-        let reversed = make_entries(&[
-            "app",
-            "appx",
-            "appabcdefghi",
-            "appabcdefghij",
-        ]);
-        let mut ordered_engine = SearchEngine::new(ordered);
-        let mut reversed_engine = SearchEngine::new(reversed);
-
-        let ordered_results = ordered_engine.search("app", 2, &empty_history(), SearchMode::Prefix);
-        let reversed_results = reversed_engine.search("app", 2, &empty_history(), SearchMode::Prefix);
-
-        let ordered_names: Vec<&str> = ordered_results.iter().map(|r| r.name.as_str()).collect();
-        let reversed_names: Vec<&str> = reversed_results.iter().map(|r| r.name.as_str()).collect();
-
-        assert_eq!(ordered_names, vec!["app", "appx"]);
-        assert_eq!(reversed_names, vec!["app", "appx"]);
-    }
-
-    #[test]
-    fn search_results_are_not_folders() {
-        let entries = make_entries(&["Firefox"]);
-        let mut engine = SearchEngine::new(entries);
-        let results = engine.search("fire", 8, &empty_history(), SearchMode::Fuzzy);
-        assert!(!results.is_empty());
-        assert!(!results[0].is_folder);
-    }
-
-    #[test]
-    fn search_prefix_mode_matches_only_prefix() {
-        let entries = make_entries(&["Notepad", "Pad Tool"]);
-        let mut engine = SearchEngine::new(entries);
-        let results = engine.search("pad", 8, &empty_history(), SearchMode::Prefix);
-        assert_eq!(results.len(), 1);
-        assert_eq!(results[0].name, "Pad Tool");
-    }
-
-    #[test]
-    fn search_substring_mode_matches_middle() {
-        let entries = make_entries(&["Visual Studio Code"]);
-        let mut engine = SearchEngine::new(entries);
-        let results = engine.search("studio", 8, &empty_history(), SearchMode::Substring);
-        assert_eq!(results.len(), 1);
-    }
-
-    #[test]
-    fn search_with_extension_matches_stem_entry() {
-        // "SSP.exe" と入力して、name="SSP", target_path="C:\\fake\\SSP.exe" にマッチする
-        let entries = vec![AppEntry {
-            name: "SSP".to_string(),
-            target_path: "C:\\fake\\SSP.exe".to_string(),
-            is_folder: false,
-        }];
-        let mut engine = SearchEngine::new(entries);
-        let results = engine.search("SSP.exe", 8, &empty_history(), SearchMode::Prefix);
-        assert_eq!(results.len(), 1);
-        assert_eq!(results[0].name, "SSP");
-    }
-
-    #[test]
-    fn search_with_extension_substring_mode() {
-        let entries = vec![AppEntry {
-            name: "SSP".to_string(),
-            target_path: "C:\\fake\\SSP.exe".to_string(),
-            is_folder: false,
-        }];
-        let mut engine = SearchEngine::new(entries);
-        let results = engine.search("ssp.exe", 8, &empty_history(), SearchMode::Substring);
-        assert_eq!(results.len(), 1);
-    }
-
-    #[test]
-    fn search_with_extension_fuzzy_mode() {
-        let entries = vec![AppEntry {
-            name: "SSP".to_string(),
-            target_path: "C:\\fake\\SSP.exe".to_string(),
-            is_folder: false,
-        }];
-        let mut engine = SearchEngine::new(entries);
-        let results = engine.search("ssp.exe", 8, &empty_history(), SearchMode::Fuzzy);
-        assert_eq!(results.len(), 1);
-    }
-
-    #[test]
-    fn search_without_extension_still_works() {
-        let entries = make_entries(&["SSP"]);
-        let mut engine = SearchEngine::new(entries);
-        let results = engine.search("SSP", 8, &empty_history(), SearchMode::Prefix);
-        assert_eq!(results.len(), 1);
-        assert_eq!(results[0].name, "SSP");
-    }
-
-    #[test]
-    fn search_with_extension_does_not_match_unrelated_exe() {
-        // "ssp.exe" で FileZilla.exe はヒットしない（stem "ssp" が fuzzy でも一致しない）
-        let entries = vec![
-            AppEntry {
-                name: "SSP".to_string(),
-                target_path: "C:\\fake\\SSP.exe".to_string(),
-                is_folder: false,
-            },
-            AppEntry {
-                name: "FileZilla".to_string(),
-                target_path: "C:\\fake\\FileZilla.exe".to_string(),
-                is_folder: false,
-            },
-        ];
-        let mut engine = SearchEngine::new(entries);
-        let results = engine.search("ssp.exe", 8, &empty_history(), SearchMode::Fuzzy);
-        assert_eq!(results.len(), 1);
-        assert_eq!(results[0].name, "SSP");
-    }
-
-    #[test]
-    fn search_with_extension_filters_by_ext() {
-        // "ssp.exe" は .lnk の SSP にはヒットしない（ファイル名 "SSP.lnk" と "ssp.exe" は不一致）
-        let entries = vec![AppEntry {
-            name: "SSP".to_string(),
-            target_path: "C:\\fake\\SSP.lnk".to_string(),
-            is_folder: false,
-        }];
-        let mut engine = SearchEngine::new(entries);
-        let results = engine.search("ssp.exe", 8, &empty_history(), SearchMode::Prefix);
-        assert!(results.is_empty());
-    }
-
-    #[test]
-    fn search_partial_ext_dot_only() {
-        // "SSP." → target_path のファイル名 "SSP.exe" に fuzzy 一致
-        let entries = vec![AppEntry {
-            name: "SSP".to_string(),
-            target_path: "C:\\fake\\SSP.exe".to_string(),
-            is_folder: false,
-        }];
-        let mut engine = SearchEngine::new(entries);
-        let results = engine.search("SSP.", 8, &empty_history(), SearchMode::Fuzzy);
-        assert_eq!(results.len(), 1);
-        assert_eq!(results[0].name, "SSP");
-    }
-
-    #[test]
-    fn search_partial_ext_dot_e() {
-        // "SSP.e" → target_path のファイル名 "SSP.exe" に fuzzy 一致
-        let entries = vec![AppEntry {
-            name: "SSP".to_string(),
-            target_path: "C:\\fake\\SSP.exe".to_string(),
-            is_folder: false,
-        }];
-        let mut engine = SearchEngine::new(entries);
-        let results = engine.search("SSP.e", 8, &empty_history(), SearchMode::Fuzzy);
-        assert_eq!(results.len(), 1);
-        assert_eq!(results[0].name, "SSP");
-    }
-
-    #[test]
-    fn search_partial_ext_dot_ex() {
-        // "SSP.ex" → target_path のファイル名 "SSP.exe" に fuzzy 一致
-        let entries = vec![AppEntry {
-            name: "SSP".to_string(),
-            target_path: "C:\\fake\\SSP.exe".to_string(),
-            is_folder: false,
-        }];
-        let mut engine = SearchEngine::new(entries);
-        let results = engine.search("SSP.ex", 8, &empty_history(), SearchMode::Fuzzy);
-        assert_eq!(results.len(), 1);
-        assert_eq!(results[0].name, "SSP");
-    }
-
-    #[test]
-    fn search_name_with_dot_matches() {
-        // name にドットを含むエントリが、ドット入りクエリでヒットする
-        let entries = vec![AppEntry {
-            name: "Dr.Web".to_string(),
-            target_path: "C:\\fake\\drweb32w.exe".to_string(),
-            is_folder: false,
-        }];
-        let mut engine = SearchEngine::new(entries);
-        let results = engine.search("Dr.Web", 8, &empty_history(), SearchMode::Fuzzy);
-        assert_eq!(results.len(), 1);
-        assert_eq!(results[0].name, "Dr.Web");
-    }
-
-    #[test]
-    fn search_name_with_dot_prefers_name() {
-        // name にドットを含むエントリが、部分一致クエリでもヒットする
-        let entries = vec![AppEntry {
-            name: "Dr.Web".to_string(),
-            target_path: "C:\\fake\\drweb32w.exe".to_string(),
-            is_folder: false,
-        }];
-        let mut engine = SearchEngine::new(entries);
-        let results = engine.search("dr.w", 8, &empty_history(), SearchMode::Fuzzy);
-        assert_eq!(results.len(), 1);
-        assert_eq!(results[0].name, "Dr.Web");
-    }
-
-    #[test]
-    fn search_double_ext_file() {
-        // 二重拡張子のファイルに対して部分一致でヒットする
-        let entries = vec![AppEntry {
-            name: "hoge".to_string(),
-            target_path: "C:\\fake\\hoge.exe.bak".to_string(),
-            is_folder: false,
-        }];
-        let mut engine = SearchEngine::new(entries);
-        let results = engine.search("hoge.exe", 8, &empty_history(), SearchMode::Fuzzy);
-        assert_eq!(results.len(), 1);
-        assert_eq!(results[0].name, "hoge");
-    }
-
-    #[test]
-    fn search_double_ext_full() {
-        // 二重拡張子のファイルに対して完全一致でヒットする
-        let entries = vec![AppEntry {
-            name: "hoge".to_string(),
-            target_path: "C:\\fake\\hoge.exe.bak".to_string(),
-            is_folder: false,
-        }];
-        let mut engine = SearchEngine::new(entries);
-        let results = engine.search("hoge.exe.bak", 8, &empty_history(), SearchMode::Fuzzy);
-        assert_eq!(results.len(), 1);
-        assert_eq!(results[0].name, "hoge");
-    }
-
-    #[test]
-    fn recent_history_empty_when_no_launches() {
-        let entries = make_entries(&["Firefox", "Chrome"]);
-        let engine = SearchEngine::new(entries);
-        let results = engine.recent_history(&empty_history(), 8);
-        assert!(results.is_empty());
-    }
-
-    #[test]
-    fn rank_cmp_breaks_full_tie_with_target_path() {
-        // index は Ord/PartialEq に不参加（tie-break キーは score→last_launched→lower_name→path）
-        let ra = ScoredEntry {
-            score: 100,
-            last_launched: 200,
-            lower_name: "tool",
-            path: "C:\\B\\tool.exe",
-            index: 0,
-        };
-        let rb = ScoredEntry {
-            score: 100,
-            last_launched: 200,
-            lower_name: "tool",
-            path: "C:\\A\\tool.exe",
-            index: 1,
-        };
-        let mut scored = [ra, rb];
-        scored.sort();
-        assert_eq!(scored[0].path, "C:\\A\\tool.exe");
-        assert_eq!(scored[1].path, "C:\\B\\tool.exe");
-    }
-
-    #[test]
-    fn has_dot_uses_cached_lower_file_name() {
-        let entries = vec![AppEntry {
-            name: "Dummy".to_string(),
-            target_path: "C:\\fake\\Tool.EXE".to_string(),
-            is_folder: false,
-        }];
-        let mut engine = SearchEngine::new(entries);
-        let results = engine.search("tool.exe", 8, &empty_history(), SearchMode::Substring);
-        assert_eq!(results.len(), 1);
-        assert_eq!(results[0].name, "Dummy");
-    }
-
-    #[test]
-    fn has_dot_handles_missing_file_name_without_panic() {
-        let entries = vec![AppEntry {
-            name: "Dummy".to_string(),
-            target_path: "C:\\".to_string(),
-            is_folder: false,
-        }];
-        let mut engine = SearchEngine::new(entries);
-        let results = engine.search("dummy.exe", 8, &empty_history(), SearchMode::Fuzzy);
-        assert!(results.is_empty());
-    }
-
-    #[test]
-    fn adjusted_history_boost_disabled_uses_raw_value() {
-        let adjusted =
-            adjusted_history_boost(SearchMode::Fuzzy, 100, 250, SearchOptions::default());
-        assert_eq!(adjusted, 250);
-    }
-
-    #[test]
-    fn adjusted_history_boost_caps_only_in_fuzzy_mode() {
-        let config = SearchOptions {
-            normalization: SearchHistoryNormalizationConfig::FuzzyRelativeCap,
-            fuzzy_history_cap_ratio: 0.30,
-            ..SearchOptions::default()
-        };
-        let adjusted = adjusted_history_boost(SearchMode::Prefix, 100, 250, config);
-        assert_eq!(adjusted, 250);
-    }
-
-    #[test]
-    fn adjusted_history_boost_caps_fuzzy_history_by_base_ratio() {
-        let config = SearchOptions {
-            normalization: SearchHistoryNormalizationConfig::FuzzyRelativeCap,
-            fuzzy_history_cap_ratio: 0.30,
-            ..SearchOptions::default()
-        };
-        let adjusted = adjusted_history_boost(SearchMode::Fuzzy, 100, 250, config);
-        assert_eq!(adjusted, 30);
-    }
-
-    #[test]
-    fn adjusted_history_boost_zeroes_when_base_is_non_positive() {
-        let config = SearchOptions {
-            normalization: SearchHistoryNormalizationConfig::FuzzyRelativeCap,
-            fuzzy_history_cap_ratio: 0.30,
-            ..SearchOptions::default()
-        };
-        let adjusted = adjusted_history_boost(SearchMode::Fuzzy, -50, 250, config);
-        assert_eq!(adjusted, 0);
-    }
-
-    #[test]
-    fn search_with_options_disabled_matches_legacy_search() {
-        let entries = make_entries(&["alpha", "alpaca", "alpine"]);
-        let mut engine = SearchEngine::new(entries);
-        let mut history = empty_history();
-        for _ in 0..50 {
-            history.record_launch("C:\\fake\\alpaca.lnk", "alp");
-        }
-
-        let legacy = engine.search("alp", 8, &history, SearchMode::Fuzzy);
-        let explicit = engine.search_with_options(
-            "alp",
-            8,
-            &history,
-            SearchMode::Fuzzy,
-            SearchOptions::default(),
-        );
-        assert_eq!(legacy, explicit);
-    }
-
-    // --- 大文字パス正規化テスト ---
-
-    #[test]
-    fn recent_history_matches_case_insensitive_path() {
-        // 大文字パスで記録した起動履歴が、元ケース AppEntry と照合できる
-        let entries = vec![AppEntry {
-            name: "App".to_string(),
-            target_path: "C:\\Fake\\App.lnk".to_string(),
-            is_folder: false,
-        }];
-        let engine = SearchEngine::new(entries);
-        let mut history = empty_history();
-        history.record_launch("C:\\FAKE\\APP.LNK", "app");
-
-        let results = engine.recent_history(&history, 8);
-        assert_eq!(results.len(), 1);
-        assert_eq!(results[0].name, "App");
-    }
-
-    #[test]
-    fn query_boost_matches_case_insensitive_path() {
-        // 大文字パスで記録したクエリ別履歴がスコアブーストに反映され、
-        // 同スコアの競合エントリより上位に来る
-        let entries = vec![
-            AppEntry {
-                name: "App".to_string(),
-                target_path: "C:\\Fake\\App.lnk".to_string(),
-                is_folder: false,
-            },
-            AppEntry {
-                name: "AppX".to_string(),
-                target_path: "C:\\Other\\appx.lnk".to_string(),
-                is_folder: false,
-            },
-        ];
-        let mut engine = SearchEngine::new(entries);
-        let mut history = empty_history();
-        for _ in 0..10 {
-            history.record_launch("C:\\FAKE\\APP.LNK", "app");
-        }
-
-        let results = engine.search_with_options(
-            "app",
-            8,
-            &history,
-            SearchMode::Prefix,
-            SearchOptions::default(),
-        );
-        assert!(!results.is_empty());
-        // 履歴ブーストにより "App" が "AppX" より上位に来る
-        assert_eq!(results[0].name, "App");
-    }
-
-    // --- char_bitmask テスト ---
-
-    #[test]
-    fn bitmask_lowercase_letters() {
-        let mask = char_bitmask("abc");
-        assert_ne!(mask & (1 << 0), 0); // a
-        assert_ne!(mask & (1 << 1), 0); // b
-        assert_ne!(mask & (1 << 2), 0); // c
-        assert_eq!(mask & (1 << 3), 0); // d not present
-    }
-
-    #[test]
-    fn bitmask_digits() {
-        let mask = char_bitmask("a1b2");
-        assert_ne!(mask & (1 << 0), 0);      // a
-        assert_ne!(mask & (1 << 1), 0);      // b
-        assert_ne!(mask & (1 << 27), 0);     // '1' = bit 26+1
-        assert_ne!(mask & (1 << 28), 0);     // '2' = bit 26+2
-    }
-
-    #[test]
-    fn bitmask_non_alnum_ignored() {
-        let mask = char_bitmask("a-b.c");
-        assert_ne!(mask & (1 << 0), 0); // a
-        assert_ne!(mask & (1 << 1), 0); // b
-        assert_ne!(mask & (1 << 2), 0); // c
-        // '-' and '.' don't set any bits
-        assert_eq!(mask.count_ones(), 3);
-    }
-
-    #[test]
-    fn bitmask_filter_skips_non_matching_entries() {
-        // "xyz" のクエリでビットマスクが "Firefox" にマッチしないことを確認
-        let entries = make_entries(&["Firefox", "Chrome", "Xyzer"]);
-        let mut engine = SearchEngine::new(entries);
-        let results = engine.search("xyz", 8, &empty_history(), SearchMode::Fuzzy);
-        // "Firefox" と "Chrome" は x,y,z の全文字を含まないのでスキップされる
-        // "Xyzer" は x,y,z を含むのでマッチ候補になる
-        assert_eq!(results.len(), 1);
-        assert_eq!(results[0].name, "Xyzer");
-    }
-
-    #[test]
-    fn bitmask_filter_does_not_skip_accented_entries() {
-        // nucleo はアクセント正規化（é→e）でマッチするため、
-        // ビットマスクフィルタがアクセント付きエントリを除外してはならない
-        let entries = vec![AppEntry {
-            name: "Café".to_string(),
-            target_path: "C:\\fake\\Café.lnk".to_string(),
-            is_folder: false,
-        }];
-        let mut engine = SearchEngine::new(entries);
-        let results = engine.search("cafe", 8, &empty_history(), SearchMode::Fuzzy);
-        assert_eq!(results.len(), 1);
-        assert_eq!(results[0].name, "Café");
-    }
-
-    // --- アクセント正規化統一テスト ---
-
-    #[test]
-    fn prefix_matches_accented_entry() {
-        let entries = vec![AppEntry {
-            name: "Café".to_string(),
-            target_path: "C:\\fake\\Café.lnk".to_string(),
-            is_folder: false,
-        }];
-        let mut engine = SearchEngine::new(entries);
-        let results = engine.search("cafe", 8, &empty_history(), SearchMode::Prefix);
-        assert_eq!(results.len(), 1);
-        assert_eq!(results[0].name, "Café");
-    }
-
-    #[test]
-    fn substring_matches_accented_entry() {
-        let entries = vec![AppEntry {
-            name: "Résumé Builder".to_string(),
-            target_path: "C:\\fake\\Résumé Builder.lnk".to_string(),
-            is_folder: false,
-        }];
-        let mut engine = SearchEngine::new(entries);
-        let results = engine.search("resume", 8, &empty_history(), SearchMode::Substring);
-        assert_eq!(results.len(), 1);
-        assert_eq!(results[0].name, "Résumé Builder");
-    }
-
-    // --- パフォーマンス計測 ---
-    // `cargo test -p snotra-core bench_ -- --ignored --nocapture` で実行
-
-    fn make_bench_entries(n: usize) -> Vec<AppEntry> {
-        // 実際のアプリ名に近い多様な文字列を生成する
-        let prefixes = [
-            "Microsoft", "Adobe", "Google", "Apple", "Mozilla", "Visual", "Windows",
-            "System", "App", "Tool", "Launcher", "Manager", "Explorer", "Editor",
-        ];
-        let suffixes = [
-            "Studio", "Reader", "Player", "Code", "Settings", "Control", "Panel",
-            "Viewer", "Browser", "Assistant", "Helper", "Updater", "Installer",
-        ];
-        (0..n)
-            .map(|i| {
-                let name = format!(
-                    "{} {} {}",
-                    prefixes[i % prefixes.len()],
-                    suffixes[i % suffixes.len()],
-                    i
-                );
-                AppEntry {
-                    target_path: format!("C:\\Program Files\\App{}\\app{}.lnk", i, i),
-                    name,
-                    is_folder: false,
-                }
-            })
-            .collect()
-    }
-
-    /// カタカナ主体のエントリ生成。migemo の主対象（かな名）で kana 実体の上界を測るため。
-    fn make_bench_entries_katakana(n: usize) -> Vec<AppEntry> {
-        let prefixes = [
-            "マイクロソフト", "アドビ", "グーグル", "アップル", "システム",
-            "ツール", "ランチャー", "マネージャー", "エクスプローラー", "エディター",
-        ];
-        let suffixes = [
-            "スタジオ", "リーダー", "プレイヤー", "コード", "セッティング",
-            "コントロール", "ビューアー", "ブラウザ", "アシスタント", "ヘルパー",
-        ];
-        (0..n)
-            .map(|i| AppEntry {
-                name: format!(
-                    "{}{}{}",
-                    prefixes[i % prefixes.len()],
-                    suffixes[i % suffixes.len()],
-                    i
-                ),
-                target_path: format!("C:\\Program Files\\App{}\\app{}.lnk", i, i),
-                is_folder: false,
-            })
-            .collect()
-    }
-
-    /// kana_lower_names（`Vec<Box<str>>`）のメモリ実体を分解計測する。
-    /// migemo 無効時はこの構造体ごと空になるため、ここで測る合計が削減量に相当する。
-    /// - ポインタ配列: `capacity * size_of::<Box<str>>()`（ファットポインタ 16B/要素）
-    /// - 文字列実体: 各 Box<str> のバイト長総和（要求バイト）
-    /// - 16B 粒度丸め: 各実体を個別ヒープ確保とみなし 16B 境界へ切り上げた実 RSS 寄りの上界
-    fn measure_kana_footprint(label: &str, entries: &[AppEntry]) {
-        let kana: Vec<Box<str>> = entries
-            .iter()
-            .map(|e| to_kana(&to_lower_folded(&e.name)).into_boxed_str())
-            .collect();
-        let n = kana.len().max(1);
-        let ptr_bytes = kana.capacity() * std::mem::size_of::<Box<str>>();
-        let content_bytes: usize = kana.iter().map(|s| s.len()).sum();
-        let rounded_content: usize = kana.iter().map(|s| s.len().max(1).div_ceil(16) * 16).sum();
-        let transformed = entries
-            .iter()
-            .filter(|e| {
-                let lf = to_lower_folded(&e.name);
-                to_kana(&lf) != lf
-            })
-            .count();
-        let requested = ptr_bytes + content_bytes;
-        let rss_ish = ptr_bytes + rounded_content;
-        let mb = |b: usize| b as f64 / (1024.0 * 1024.0);
-        println!(
-            "[{label}] n={n} かな変換割合={pct:.0}%\n  \
-             ポインタ {ptr_bytes}B ({:.2}MB) + 実体 {content_bytes}B ({:.2}MB) \
-             = 要求 {requested}B ({:.2}MB) | 16B丸め {rss_ish}B ({:.2}MB)\n  \
-             1件あたり 要求 {}B / 丸め {}B",
-            mb(ptr_bytes),
-            mb(content_bytes),
-            mb(requested),
-            mb(rss_ish),
-            requested / n,
-            rss_ish / n,
-            pct = transformed as f64 * 100.0 / n as f64,
-        );
-    }
-
-    #[test]
-    #[ignore]
-    fn measure_kana_memory_footprint() {
-        for &n in &[1_000usize, 10_000, 50_000, 100_000] {
-            measure_kana_footprint(&format!("ascii    n={n}"), &make_bench_entries(n));
-            measure_kana_footprint(&format!("katakana n={n}"), &make_bench_entries_katakana(n));
-        }
-    }
-
-    /// lower_names を name と共有する `enum LowerName`（issue #336）の ROI を実測する。
-    /// `Same`（`to_lower_folded(name) == name`）の比率と、Same エントリで排除できる
-    /// ヒープ文字列実体（現状 `Box<str>` で name とは別アロケーション）の削減量を分解計測する。
-    /// - 16B 粒度丸め: 個別ヒープ確保とみなし 16B 境界へ切り上げた実 RSS 寄りの上界
-    fn measure_lower_name_footprint(label: &str, entries: &[AppEntry]) {
-        let n = entries.len().max(1);
-        let mut same = 0usize;
-        let mut saved_requested = 0usize;
-        let mut saved_rounded = 0usize;
-        for e in entries {
-            let lower = to_lower_folded(&e.name);
-            if lower == e.name {
-                same += 1;
-                let bytes = lower.len();
-                saved_requested += bytes;
-                saved_rounded += bytes.max(1).div_ceil(16) * 16;
-            }
-        }
-        let mb = |b: usize| b as f64 / (1024.0 * 1024.0);
-        let pct = same as f64 * 100.0 / n as f64;
-        println!(
-            "[{label}] n={n} Same={same} ({pct:.1}%)\n  \
-             削減（Same のヒープ文字列実体）要求 {saved_requested}B ({:.3}MB) | 16B丸め {saved_rounded}B ({:.3}MB)\n  \
-             全entry平均 要求 {}B/件 / 丸め {}B/件 | 50k換算 丸め {:.3}MB",
-            mb(saved_requested),
-            mb(saved_rounded),
-            saved_requested / n,
-            saved_rounded / n,
-            mb(saved_rounded * 50_000 / n),
-        );
-    }
-
-    #[test]
-    #[ignore]
-    fn measure_lower_name_footprint_report() {
-        use crate::config::Config;
-        use crate::indexer::{scan_all, scan_path_env};
-
-        // 本番相当: 既定 scan paths（Start Menu + Desktop の .lnk）+ PATH 実行ファイル。
-        let mut real = scan_all(&Config::default_scan_paths(), false);
-        let path_entries = scan_path_env(&real, false);
-        real.extend(path_entries);
-        measure_lower_name_footprint("real (start menu + desktop + PATH)", &real);
-
-        // 参考: 合成ベンチ（名前パターンが本番非代表＝Same はほぼ出ない）。
-        measure_lower_name_footprint("synthetic ascii    n=10000", &make_bench_entries(10_000));
-        measure_lower_name_footprint("synthetic katakana n=10000", &make_bench_entries_katakana(10_000));
-    }
-
-    fn bench_search(label: &str, n: usize, queries: &[&str]) {
-        use std::time::Instant;
-        let entries = make_bench_entries(n);
-        let mut engine = SearchEngine::new(entries);
-        let history = empty_history();
-
-        // ウォームアップ（rayon スレッドプールの初期化を除外）
-        for q in queries {
-            let _ = engine.search(q, 10, &history, SearchMode::Fuzzy);
-        }
-
-        let iters = 20usize;
-        let mut total_ns = 0u128;
-        for _ in 0..iters {
-            for q in queries {
-                let t = Instant::now();
-                let _ = engine.search(q, 10, &history, SearchMode::Fuzzy);
-                total_ns += t.elapsed().as_nanos();
-            }
-        }
-
-        let avg_us = total_ns / (iters * queries.len()) as u128 / 1000;
-        println!("[{label}] entries={n}, avg={avg_us}µs ({} queries × {iters} iters)", queries.len());
-    }
-
-    #[test]
-    #[ignore]
-    fn bench_fuzzy_search_scaling() {
-        let queries = ["vis", "code", "micro", "app", "sett"];
-        for &n in &[1_000, 10_000, 50_000, 100_000, 300_000] {
-            bench_search("fuzzy", n, &queries);
-        }
-    }
-
-    #[test]
-    #[ignore]
-    fn bench_fuzzy_path_query_scaling() {
-        // パス区切り含有クエリ: has_path_sep=true でビットマスク pre-filter がスキップされる
-        let queries = ["fake\\vis", "fake\\code", "fake\\micro"];
-        for &n in &[1_000, 10_000, 50_000, 100_000, 300_000] {
-            bench_search("fuzzy_path", n, &queries);
-        }
-    }
-
-    fn bench_new(label: &str, n: usize) {
-        use std::time::Instant;
-        let entries = make_bench_entries(n);
-
-        // rayon スレッドプール初期化を計測から除外するためのウォームアップ
-        let _ = SearchEngine::new(entries.clone());
-
-        // 構築自体にかかる時間のみを計測する。
-        // entries.clone() のコストは Vec<AppEntry> の単純コピーであり、
-        // new() 内の lower_fold・ビットマスク計算・normalized_keys に比べ微小。
-        let iters = 5usize;
-        let mut total_ns = 0u128;
-        for _ in 0..iters {
-            let cloned = entries.clone();
-            let t = Instant::now();
-            let _ = SearchEngine::new(cloned);
-            total_ns += t.elapsed().as_nanos();
-        }
-
-        let avg_ms = total_ns / iters as u128 / 1_000_000;
-        println!("[{label}] entries={n}, avg={avg_ms}ms ({iters} iters)");
-    }
-
-    #[test]
-    #[ignore]
-    fn bench_new_scaling() {
-        for &n in &[1_000, 10_000, 50_000, 100_000, 300_000] {
-            bench_new("new", n);
-        }
-    }
-
-    /// migemo on/off の構築コスト差を計測する（issue #337）。
-    /// kana 構築（to_kana の全エントリ map）をスキップした分の差を可視化する。
-    /// 構築はロック外（PrebuiltIndex::new）で行われ、ロック保持時間は
-    /// apply_prebuilt_index の O(1) ムーブのみ（migemo 状態に依存しない）。
-    fn bench_new_migemo(label: &str, n: usize, migemo_enabled: bool) {
-        use std::time::Instant;
-        let entries = make_bench_entries_katakana(n);
-        let _ = SearchEngine::new_with_migemo(entries.clone(), migemo_enabled); // warmup
-        let iters = 5usize;
-        let mut total_ns = 0u128;
-        for _ in 0..iters {
-            let cloned = entries.clone();
-            let t = Instant::now();
-            let _ = SearchEngine::new_with_migemo(cloned, migemo_enabled);
-            total_ns += t.elapsed().as_nanos();
-        }
-        let avg_us = total_ns / iters as u128 / 1000;
-        println!("[{label}] entries={n}, migemo={migemo_enabled}, avg={avg_us}µs ({iters} iters)");
-    }
-
-    #[test]
-    #[ignore]
-    fn bench_new_migemo_on_off() {
-        for &n in &[1_000, 10_000, 50_000, 100_000] {
-            bench_new_migemo("new_migemo_on ", n, true);
-            bench_new_migemo("new_migemo_off", n, false);
-        }
-    }
-
-    #[test]
-    fn history_boost_unified_across_accent_variants() {
-        // "résumé" で起動記録 → "resume" で検索時に履歴ブーストが効く
-        let entries = vec![
-            AppEntry {
-                name: "Résumé Builder".to_string(),
-                target_path: "C:\\fake\\Résumé Builder.lnk".to_string(),
-                is_folder: false,
-            },
-            AppEntry {
-                name: "Resume Helper".to_string(),
-                target_path: "C:\\fake\\Resume Helper.lnk".to_string(),
-                is_folder: false,
-            },
-        ];
-        let mut engine = SearchEngine::new(entries);
-        let mut history = empty_history();
-        // "résumé" で Résumé Builder を多数起動
-        for _ in 0..20 {
-            history.record_launch("C:\\fake\\Résumé Builder.lnk", "résumé");
-        }
-        // "resume"（アクセントなし）で検索 → 履歴ブーストが効いて Résumé Builder が上位
-        let results = engine.search("resume", 8, &history, SearchMode::Fuzzy);
-        assert!(!results.is_empty());
-        assert_eq!(results[0].name, "Résumé Builder");
-    }
-
-    // --- インクリメンタル検索テスト ---
-
-    #[test]
-    fn incremental_search_gives_correct_results_on_extension() {
-        let names = &["Firefox", "Final Cut", "Chrome", "Finder", "Fire TV"];
-        let mut engine = SearchEngine::new(make_entries(names));
-        let h = empty_history();
-
-        // 連続する monotonic 拡張で incremental パスを使う
-        let _ = engine.search("fi", 8, &h, SearchMode::Fuzzy);
-        let _ = engine.search("fir", 8, &h, SearchMode::Fuzzy);
-        let incremental = engine.search("fire", 8, &h, SearchMode::Fuzzy);
-
-        // 新鮮なエンジンでの結果と一致するか確認
-        let mut fresh = SearchEngine::new(make_entries(names));
-        let fresh_result = fresh.search("fire", 8, &h, SearchMode::Fuzzy);
-
-        let inc_names: Vec<&str> = incremental.iter().map(|r| r.name.as_str()).collect();
-        let fresh_names: Vec<&str> = fresh_result.iter().map(|r| r.name.as_str()).collect();
-        assert_eq!(inc_names, fresh_names);
-    }
-
-    #[test]
-    fn incremental_search_fallback_on_backspace() {
-        // "Firm" は fuzzy "fir" にマッチするが "fire" にはマッチしない（'e' がない）。
-        // "fire" → "fir" はバックスペースなので incremental パスは使えない。
-        // full scan にフォールバックしていなければ "Firm" が結果から漏れる。
-        let mut engine =
-            SearchEngine::new(make_entries(&["Firefox", "Final Cut", "Chrome", "Firm"]));
-        let h = empty_history();
-
-        let _ = engine.search("fire", 8, &h, SearchMode::Fuzzy);
-        // "fir" は "fire" の拡張ではない → full scan にフォールバック
-        let results = engine.search("fir", 8, &h, SearchMode::Fuzzy);
-
-        let names: Vec<&str> = results.iter().map(|r| r.name.as_str()).collect();
-        assert!(names.contains(&"Firefox"), "full scan で Firefox が返る必要がある");
-        assert!(
-            names.contains(&"Firm"),
-            "Firm は 'fire' にマッチしないため前回の candidates に不在。full scan でのみ検出可能"
-        );
-    }
-
-    #[test]
-    fn incremental_search_fallback_on_mode_change() {
-        let mut engine = SearchEngine::new(make_entries(&["Firefox", "Final Cut", "Chrome"]));
-        let h = empty_history();
-
-        let _ = engine.search("fi", 8, &h, SearchMode::Fuzzy);
-        // 同じクエリでもモード変更 → full scan
-        let results = engine.search("fi", 8, &h, SearchMode::Prefix);
-
-        // Prefix "fi": "firefox" / "final cut" は先頭一致、"chrome" は不一致
-        let names: Vec<&str> = results.iter().map(|r| r.name.as_str()).collect();
-        assert!(names.contains(&"Firefox"), "Prefix 'fi' で Firefox が返る必要がある");
-        assert!(names.contains(&"Final Cut"), "Prefix 'fi' で Final Cut が返る必要がある");
-        assert!(
-            !names.contains(&"Chrome"),
-            "Chrome は 'fi' で始まらないため除外される必要がある"
-        );
-    }
-
-    #[test]
-    fn incremental_search_dot_to_dot_uses_incremental() {
-        // dot → dot の拡張は incremental パスを使用できる（no-dot→dot ガードを通過する）。
-        // "ssp." → "ssp.e" はどちらもドットを含む単調拡張。
-        let entries = vec![
-            AppEntry {
-                name: "SSP".to_string(),
-                target_path: "C:\\fake\\SSP.exe".to_string(),
-                is_folder: false,
-            },
-            AppEntry {
-                name: "AnotherApp".to_string(),
-                target_path: "C:\\fake\\ssp.data".to_string(),
-                is_folder: false,
-            },
-        ];
-        let mut engine = SearchEngine::new(entries.clone());
-        let h = empty_history();
-
-        // "ssp." で両エントリが候補にキャッシュされる（prev_candidates に両インデックスが入る）
-        let _ = engine.search("ssp.", 8, &h, SearchMode::Fuzzy);
-        // "ssp.e" はドットあり拡張 → incremental で prev_candidates を再利用
-        let incremental = engine.search("ssp.e", 8, &h, SearchMode::Fuzzy);
-
-        // fresh エンジンとの比較で正確性を担保
-        let mut fresh = SearchEngine::new(entries);
-        let fresh_result = fresh.search("ssp.e", 8, &h, SearchMode::Fuzzy);
-
-        let inc_names: Vec<&str> = incremental.iter().map(|r| r.name.as_str()).collect();
-        let fresh_names: Vec<&str> = fresh_result.iter().map(|r| r.name.as_str()).collect();
-        assert_eq!(
-            inc_names,
-            fresh_names,
-            "dot→dot incremental は fresh 結果と一致する必要がある"
-        );
-    }
-
-    #[test]
-    fn incremental_search_no_dot_to_dot_falls_back_to_full_scan() {
-        // "AnotherApp" は名前では "ssp" にマッチしないが、
-        // file_name "ssp.data" は "ssp." クエリにマッチする
-        let entries = vec![
-            AppEntry {
-                name: "SSP".to_string(),
-                target_path: "C:\\fake\\SSP.exe".to_string(),
-                is_folder: false,
-            },
-            AppEntry {
-                name: "AnotherApp".to_string(),
-                target_path: "C:\\fake\\ssp.data".to_string(),
-                is_folder: false,
-            },
-        ];
-        let mut engine = SearchEngine::new(entries);
-        let h = empty_history();
-
-        // "ssp"（ドットなし）→ prev_candidates には名前マッチした SSP だけが入る
-        let _ = engine.search("ssp", 8, &h, SearchMode::Fuzzy);
-
-        // "ssp."（ドットあり）→ no-dot→dot ガードにより full scan
-        // AnotherApp の file_name "ssp.data" が "ssp." にマッチするはず
-        let results = engine.search("ssp.", 8, &h, SearchMode::Fuzzy);
-        let names: Vec<&str> = results.iter().map(|r| r.name.as_str()).collect();
-        assert!(names.contains(&"SSP"), "SSP.exe が ssp. にマッチするはず");
-        assert!(
-            names.contains(&"AnotherApp"),
-            "AnotherApp の file_name ssp.data が ssp. にマッチするはず（full scan 必須）"
-        );
-    }
-
-    #[test]
-    fn incremental_search_empty_prev_candidates_falls_back() {
-        let mut engine = SearchEngine::new(make_entries(&["Firefox", "Chrome"]));
-        let h = empty_history();
-
-        // マッチなし → prev_candidates が空になる
-        let r1 = engine.search("xyz", 8, &h, SearchMode::Fuzzy);
-        assert!(r1.is_empty());
-
-        // "xyzw" は "xyz" の拡張だが prev_candidates 空 → full scan
-        let r2 = engine.search("xyzw", 8, &h, SearchMode::Fuzzy);
-        assert!(r2.is_empty());
-
-        // キャッシュが壊れていなければ通常クエリも機能する
-        let r3 = engine.search("fire", 8, &h, SearchMode::Fuzzy);
-        assert!(!r3.is_empty());
-        assert_eq!(r3[0].name, "Firefox");
-    }
-
-    // --- migemo（ローマ字→かな）検索テスト ---
-
-    fn migemo_config() -> SearchOptions {
-        SearchOptions {
-            migemo_enabled: true,
-            migemo_min_chars: 2,
-            ..SearchOptions::default()
-        }
-    }
-
-    #[test]
-    fn kana_search_disabled_by_default() {
-        // migemo_enabled=false（デフォルト）では "dokyu" で "ドキュメント" がヒットしない
-        let entries = make_entries(&["ドキュメント", "Documents"]);
-        let mut engine = SearchEngine::new(entries);
-        let h = empty_history();
-        let results = engine.search_with_options(
-            "dokyu",
-            8,
-            &h,
-            SearchMode::Substring,
-            SearchOptions::default(), // migemo_enabled=false
-        );
-        assert!(
-            results.is_empty(),
-            "migemo 無効時は 'dokyu' でカタカナ名がヒットしてはならない"
-        );
-    }
-
-    #[test]
-    fn kana_search_matches_katakana_entry() {
-        // "dokyu" で "ドキュメント" がヒットする
-        let entries = make_entries(&["ドキュメント", "Documents"]);
-        let mut engine = SearchEngine::new(entries);
-        let h = empty_history();
-        let results = engine.search_with_options(
-            "dokyu",
-            8,
-            &h,
-            SearchMode::Substring,
-            migemo_config(),
-        );
-        let names: Vec<&str> = results.iter().map(|r| r.name.as_str()).collect();
-        assert!(
-            names.contains(&"ドキュメント"),
-            "'dokyu' で ドキュメント がヒットするはず: {:?}",
-            names
-        );
-    }
-
-    #[test]
-    fn fuzzy_kana_match_skips_latin_bitmask_prefilter() {
-        // "chatto" と "tyatto" は kana に正規化するとどちらも「ちゃっと」になる。
-        // 直接のラテン文字集合は異なるため、Fuzzy の bitmask pre-filter は
-        // kana_query がある経路を早期棄却してはならない。
-        let mut engine = SearchEngine::new(make_entries(&["tyatto"]));
-        let plan = prepare_query_plan("chatto", SearchMode::Fuzzy, &migemo_config())
-            .expect("非空クエリは QueryPlan を生成する");
-        let kana_query_mask = plan
-            .kana_query_mask
-            .expect("完全に kana へ変換できるクエリは kana mask を持つ");
-
-        assert_ne!(
-            plan.query_mask & engine.char_masks[0],
-            plan.query_mask,
-            "tyatto はラテン文字 mask では chatto を満たさない"
-        );
-        assert_eq!(
-            kana_query_mask & engine.kana_char_masks[0],
-            kana_query_mask,
-            "kana mask は chatto と tyatto の共通 kana を通す"
-        );
-        let results = engine.search_with_options(
-            "chatto",
-            8,
-            &empty_history(),
-            SearchMode::Fuzzy,
-            migemo_config(),
-        );
-
-        assert_eq!(results.len(), 1, "kana 経由で tyatto がヒットするはず");
-        assert_eq!(results[0].name, "tyatto");
-    }
-
-    #[test]
-    fn kana_search_no_false_positive_for_ascii_entry() {
-        // "dokyu" で "Documents" はヒットしない
-        let entries = make_entries(&["ドキュメント", "Documents"]);
-        let mut engine = SearchEngine::new(entries);
-        let h = empty_history();
-        let results = engine.search_with_options(
-            "dokyu",
-            8,
-            &h,
-            SearchMode::Substring,
-            migemo_config(),
-        );
-        let names: Vec<&str> = results.iter().map(|r| r.name.as_str()).collect();
-        assert!(
-            !names.contains(&"Documents"),
-            "'dokyu' で ASCII エントリ Documents がヒットしてはならない: {:?}",
-            names
-        );
-    }
-
-    #[test]
-    fn kana_search_direct_match_ranks_above_kana_match() {
-        // 直接 Substring マッチ(5000) が kana マッチ(4500) より常に上位に来ることを確認する。
-        // "どきゅめんと"（ひらがな名）を含むエントリに対して:
-        //   - "どきゅ"（直接クエリ）→ Substring スコア 5000
-        //   - "dokyu"（kana クエリ）→ kana_substring_score 4500
-        // 直接クエリの方が kana クエリより高スコアになる。
-        let entries = vec![
-            AppEntry {
-                name: "どきゅめんと".to_string(),
-                target_path: "C:\\fake\\a.lnk".to_string(),
-                is_folder: false,
-            },
-            // kana マッチ専用エントリ（直接クエリでは name がひらがなのみなのでマッチしない）
-            AppEntry {
-                name: "ドキュメント".to_string(),
-                target_path: "C:\\fake\\b.lnk".to_string(),
-                is_folder: false,
-            },
-        ];
-        let mut engine = SearchEngine::new(entries);
-        let h = empty_history();
-
-        // "どきゅめんと" に対して直接 Substring マッチしたスコアは 5000（先頭一致）
-        let direct = engine.search_with_options(
-            "どきゅ",
-            8,
-            &h,
-            SearchMode::Substring,
-            migemo_config(),
-        );
-
-        // "dokyu" → kana_query = "どきゅ" → kana_substring_score = 4500
-        let kana = engine.search_with_options(
-            "dokyu",
-            8,
-            &h,
-            SearchMode::Substring,
-            migemo_config(),
-        );
-
-        // 両方の結果が存在することを確認
-        assert!(!direct.is_empty(), "直接クエリで どきゅめんと がヒットするはず");
-        assert!(!kana.is_empty(), "kana クエリで どきゅめんと / ドキュメント がヒットするはず");
-
-        // 直接クエリで "どきゅめんと" がトップに来る（Substring の先頭一致 = 高スコア）
-        assert_eq!(
-            direct[0].name, "どきゅめんと",
-            "直接クエリ先頭一致は どきゅめんと が最高スコア"
-        );
-
-        // kana マッチのスコア(4500) < 直接マッチのスコア(5000) の確認:
-        // 同じエントリが両クエリでヒットするとき、直接クエリ側の方が先頭になる。
-        // "どきゅめんと" は直接 Substring(5000) でも kana_substring(4500) でもヒットするが、
-        // 直接クエリ時は primary_score が Some なので kana_score は使われない（OR 関係）。
-        // よって primary_score（5000）> kana_score（4500）の順序不変条件が成立する。
-        let direct_names: Vec<&str> = direct.iter().map(|r| r.name.as_str()).collect();
-        let kana_names: Vec<&str> = kana.iter().map(|r| r.name.as_str()).collect();
-        assert!(
-            direct_names.contains(&"どきゅめんと"),
-            "直接クエリで どきゅめんと がヒットするはず: {:?}",
-            direct_names
-        );
-        assert!(
-            kana_names.contains(&"どきゅめんと") || kana_names.contains(&"ドキュメント"),
-            "kana クエリで少なくとも一方がヒットするはず: {:?}",
-            kana_names
-        );
-    }
-
-    #[test]
-    fn kana_search_min_chars_blocks_short_query() {
-        // min_chars=2 のとき 1文字クエリ "a" → "あ" はヒットしない
-        let entries = make_entries(&["あいうえお"]);
-        let mut engine = SearchEngine::new(entries);
-        let h = empty_history();
-        let config = SearchOptions {
-            migemo_enabled: true,
-            migemo_min_chars: 2,
-            ..SearchOptions::default()
-        };
-        let results = engine.search_with_options("a", 8, &h, SearchMode::Substring, config);
-        assert!(
-            results.is_empty(),
-            "1文字クエリは migemo_min_chars=2 でブロックされるはず"
-        );
-    }
-
-    #[test]
-    fn kana_search_partial_romaji_not_used() {
-        // "dok" → "どk"（'k' が残る） → kana_query が None になりヒットしない
-        let entries = make_entries(&["ドキュメント"]);
-        let mut engine = SearchEngine::new(entries);
-        let h = empty_history();
-        let results = engine.search_with_options(
-            "dok",
-            8,
-            &h,
-            SearchMode::Substring,
-            migemo_config(),
-        );
-        // "dok" は部分ローマ字で変換後に 'k' が残るため kana_query=None
-        // ドキュメント は Substring("dok") にも直接マッチしないため空
-        assert!(
-            results.is_empty(),
-            "部分ローマ字 'dok' では kana マッチもしないはず: {:?}",
-            results.iter().map(|r| r.name.as_str()).collect::<Vec<_>>()
-        );
-    }
-
-    // --- incremental + kana 遷移テスト ---
-
-    #[test]
-    fn incremental_kana_min_chars_crossing_does_not_lose_kana_match() {
-        // "d" (min_chars=2 未満 → kana_query=None) → "do" (min_chars=2 到達 → kana_query=Some("ど"))
-        // 前回の prev_candidates には kana マッチが含まれていない。
-        // full scan にフォールバックしなければ "ドキュメント" が欠落する。
-        let entries = make_entries(&["ドキュメント", "Discord"]);
-        let mut engine = SearchEngine::new(entries);
-        let h = empty_history();
-        let cfg = migemo_config();
-
-        // "d" → kana_query=None（min_chars=2 未満）
-        let _ = engine.search_with_options("d", 8, &h, SearchMode::Substring, cfg);
-
-        // "do" → kana_query=Some("ど")。前回 prev_kana_query=None なので full scan。
-        let results = engine.search_with_options("do", 8, &h, SearchMode::Substring, cfg);
-        let names: Vec<&str> = results.iter().map(|r| r.name.as_str()).collect();
-        assert!(
-            names.contains(&"ドキュメント"),
-            "min_chars 超え遷移で ドキュメント がヒットするはず（full scan フォールバック必須）: {:?}",
-            names
-        );
-    }
-
-    #[test]
-    fn incremental_kana_ascii_residue_clearing_does_not_lose_kana_match() {
-        // "dok" (ASCII 残留 'k' → kana_query=None) → "dokyu" (完全変換 → kana_query=Some("どきゅ"))
-        // prev_candidates に kana マッチが含まれていないため full scan が必要。
-        let entries = make_entries(&["ドキュメント", "Docking Station"]);
-        let mut engine = SearchEngine::new(entries);
-        let h = empty_history();
-        let cfg = migemo_config();
-
-        // "dok" → kana_query=None（"どk" に ASCII 残留）
-        let _ = engine.search_with_options("dok", 8, &h, SearchMode::Substring, cfg);
-
-        // "dokyu" → kana_query=Some("どきゅ")。prev_kana_query=None なので full scan。
-        let results = engine.search_with_options("dokyu", 8, &h, SearchMode::Substring, cfg);
-        let names: Vec<&str> = results.iter().map(|r| r.name.as_str()).collect();
-        assert!(
-            names.contains(&"ドキュメント"),
-            "ASCII 残留解消遷移で ドキュメント がヒットするはず（full scan フォールバック必須）: {:?}",
-            names
-        );
-    }
-
-    #[test]
-    fn incremental_kana_to_kana_reuses_cache() {
-        // "do" (kana_query=Some) → "doku" (kana_query=Some) は incremental で正しく動作する。
-        // kana→kana の拡張なので prev_candidates を再利用して問題ない。
-        let entries = make_entries(&["ドキュメント", "Discord", "Chrome"]);
-        let mut engine = SearchEngine::new(entries);
-        let h = empty_history();
-        let cfg = migemo_config();
-
-        let _ = engine.search_with_options("do", 8, &h, SearchMode::Substring, cfg);
-        let incremental =
-            engine.search_with_options("doku", 8, &h, SearchMode::Substring, cfg);
-
-        // fresh エンジンとの比較で正確性を担保
-        let mut fresh = SearchEngine::new(make_entries(&["ドキュメント", "Discord", "Chrome"]));
-        let fresh_result =
-            fresh.search_with_options("doku", 8, &h, SearchMode::Substring, cfg);
-
-        let inc_names: Vec<&str> = incremental.iter().map(|r| r.name.as_str()).collect();
-        let fresh_names: Vec<&str> = fresh_result.iter().map(|r| r.name.as_str()).collect();
-        assert_eq!(
-            inc_names, fresh_names,
-            "kana→kana incremental は fresh 結果と一致する必要がある"
-        );
-    }
-
-    #[test]
-    fn incremental_kana_non_monotonic_n_vowel_falls_back() {
-        // "kan" → kana="かん", "kana" → kana="かな"
-        // ローマ字→かな変換は非単調: 末尾の「ん」が「な」に変わる。
-        // "かな" は "かん" の prefix 拡張ではないため incremental は使えない。
-        // "かなめ" のようなエントリは "かん" でヒットしないが "かな" でヒットする。
-        let entries = make_entries(&["かなめ", "かんな", "Kanata"]);
-        let mut engine = SearchEngine::new(entries);
-        let h = empty_history();
-        let cfg = migemo_config();
-
-        // "kan" → kana_query=Some("かん")
-        let r1 = engine.search_with_options("kan", 8, &h, SearchMode::Substring, cfg);
-        let names1: Vec<&str> = r1.iter().map(|r| r.name.as_str()).collect();
-        assert!(
-            names1.contains(&"かんな"),
-            "\"kan\"(かん) で「かんな」がヒットするはず: {:?}",
-            names1
-        );
-
-        // "kana" → kana_query=Some("かな")。"かな" は "かん" の prefix ではない → full scan。
-        let r2 = engine.search_with_options("kana", 8, &h, SearchMode::Substring, cfg);
-        let names2: Vec<&str> = r2.iter().map(|r| r.name.as_str()).collect();
-        assert!(
-            names2.contains(&"かなめ"),
-            "非単調遷移 kan→kana で「かなめ」がヒットするはず（full scan フォールバック必須）: {:?}",
-            names2
-        );
-    }
-
-    #[test]
-    fn incremental_kana_monotonic_extension_reuses_cache() {
-        // "ka" → kana="か", "kan" → kana="かん"
-        // "かん" は "か" の prefix 拡張 → incremental が使える。
-        let entries = make_entries(&["かなめ", "かんな", "Chrome"]);
-        let mut engine = SearchEngine::new(entries);
-        let h = empty_history();
-        let cfg = migemo_config();
-
-        let _ = engine.search_with_options("ka", 8, &h, SearchMode::Substring, cfg);
-        let incremental =
-            engine.search_with_options("kan", 8, &h, SearchMode::Substring, cfg);
-
-        let mut fresh = SearchEngine::new(make_entries(&["かなめ", "かんな", "Chrome"]));
-        let fresh_result =
-            fresh.search_with_options("kan", 8, &h, SearchMode::Substring, cfg);
-
-        let inc_names: Vec<&str> = incremental.iter().map(|r| r.name.as_str()).collect();
-        let fresh_names: Vec<&str> = fresh_result.iter().map(|r| r.name.as_str()).collect();
-        assert_eq!(
-            inc_names, fresh_names,
-            "か→かん は単調拡張なので incremental は fresh と一致するはず"
-        );
-    }
-
-    // --- migemo 条件付き構築テスト (issue #337) ---
-
-    #[test]
-    fn migemo_disabled_build_leaves_kana_empty() {
-        // migemo OFF で構築 → kana 未構築。その後 migemo ON で検索しても
-        // panic せず、kana マッチは出ない（空ガード）。
-        let entries = make_entries(&["ドキュメント", "Documents"]);
-        let mut engine = SearchEngine::new_with_migemo(entries, false);
-        assert!(engine.kana_lower_names.is_empty());
-        assert!(engine.kana_char_masks.is_empty());
-        let results = engine.search_with_options(
-            "dokyu",
-            8,
-            &empty_history(),
-            SearchMode::Substring,
-            migemo_config(),
-        );
-        assert!(
-            results.is_empty(),
-            "kana 未構築時は migemo 検索が空（panic せず）: {:?}",
-            results.iter().map(|r| r.name.as_str()).collect::<Vec<_>>()
-        );
-    }
-
-    #[test]
-    fn migemo_enabled_build_matches_kana() {
-        // migemo ON で構築 → kana 構築済み → ローマ字検索がヒット
-        let entries = make_entries(&["ドキュメント", "Documents"]);
-        let mut engine = SearchEngine::new_with_migemo(entries, true);
-        assert_eq!(engine.kana_lower_names.len(), engine.entries.len());
-        assert_eq!(engine.kana_char_masks.len(), engine.entries.len());
-        let results = engine.search_with_options(
-            "dokyu",
-            8,
-            &empty_history(),
-            SearchMode::Substring,
-            migemo_config(),
-        );
-        let names: Vec<&str> = results.iter().map(|r| r.name.as_str()).collect();
-        assert!(
-            names.contains(&"ドキュメント"),
-            "migemo ON 構築で ドキュメント がヒット: {:?}",
-            names
-        );
-    }
-
-    #[test]
-    fn kana_index_follows_migemo_on_off_on() {
-        // on→off→on の各構築で kana 挙動が追従する（必須条件 #3）
-        let names = &["ドキュメント"];
-        fn romaji_hits(engine: &mut SearchEngine) -> bool {
-            !engine
-                .search_with_options(
-                    "dokyu",
-                    8,
-                    &empty_history(),
-                    SearchMode::Substring,
-                    SearchOptions {
-                        migemo_enabled: true,
-                        migemo_min_chars: 2,
-                        ..SearchOptions::default()
-                    },
-                )
-                .is_empty()
-        }
-        let mut on1 = SearchEngine::new_with_migemo(make_entries(names), true);
-        assert!(romaji_hits(&mut on1), "on(1): kana ヒットするはず");
-        let mut off = SearchEngine::new_with_migemo(make_entries(names), false);
-        assert!(!romaji_hits(&mut off), "off: kana 未構築でヒットしない");
-        let mut on2 = SearchEngine::new_with_migemo(make_entries(names), true);
-        assert!(romaji_hits(&mut on2), "on(2): 再構築で kana 復活");
-    }
-
-    #[test]
-    fn incremental_with_kana_disabled_build_no_panic() {
-        // kana 空（migemo off 構築）+ 検索時 migemo on + incremental シーケンスでも
-        // panic せず、kana-off fresh エンジンと一致する（cache-check 由来の追加テスト）。
-        let names = &["ドキュメント", "Discord", "Chrome"];
-        let cfg = migemo_config();
-        let h = empty_history();
-        let mut engine = SearchEngine::new_with_migemo(make_entries(names), false);
-        let _ = engine.search_with_options("do", 8, &h, SearchMode::Substring, cfg);
-        let inc = engine.search_with_options("doku", 8, &h, SearchMode::Substring, cfg);
-
-        let mut fresh = SearchEngine::new_with_migemo(make_entries(names), false);
-        let fresh_r = fresh.search_with_options("doku", 8, &h, SearchMode::Substring, cfg);
-
-        let inc_names: Vec<&str> = inc.iter().map(|r| r.name.as_str()).collect();
-        let fresh_names: Vec<&str> = fresh_r.iter().map(|r| r.name.as_str()).collect();
-        assert_eq!(inc_names, fresh_names, "kana 空 + incremental は fresh と一致");
-    }
-
-    // --- パスマッチング テスト ---
-
-    fn make_entry(name: &str, path: &str) -> AppEntry {
-        AppEntry {
-            name: name.to_string(),
-            target_path: path.to_string(),
-            is_folder: false,
-        }
-    }
-
-    #[test]
-    fn path_match_substring_finds_entry_by_path_segment() {
-        let entries = vec![make_entry("app", "C:\\tool\\editor\\app.exe")];
-        let mut engine = SearchEngine::new(entries);
-        let results = engine.search("tool\\editor", 8, &empty_history(), SearchMode::Substring);
-        assert_eq!(results.len(), 1);
-        assert_eq!(results[0].name, "app");
-    }
-
-    #[test]
-    fn path_match_score_below_name_match() {
-        // "editor" → name="editor" に Substring マッチ (score 5000系)
-        //         → name="app" はマッチしない（path にも "editor" を含むがクエリにパス区切りなし）
-        // パス区切りなしのクエリではパスマッチは試行されない
-        let entries = vec![
-            make_entry("editor", "C:\\tool\\editor\\editor.exe"),
-            make_entry("app", "C:\\tool\\editor\\app.exe"),
-        ];
-        let mut engine = SearchEngine::new(entries);
-        let results = engine.search("editor", 8, &empty_history(), SearchMode::Substring);
-        // "editor" は name マッチ、"app" は name にもパスにも（パス区切りなしで試行されない）マッチしない
-        assert_eq!(results.len(), 1);
-        assert_eq!(results[0].name, "editor");
-
-        // パス区切りありのクエリで比較: "tool\\editor" は両方のパスにマッチするが、
-        // "editor" は name にも Substring マッチ → name_score(5000系) > path_score(3000系)
-        let entries2 = vec![
-            make_entry("editor", "C:\\tool\\editor\\editor.exe"),
-            make_entry("app", "C:\\tool\\editor\\app.exe"),
-        ];
-        let mut engine2 = SearchEngine::new(entries2);
-        let results2 =
-            engine2.search("tool\\editor", 8, &empty_history(), SearchMode::Substring);
-        assert_eq!(results2.len(), 2);
-        // "editor" の lower_name は "editor"。クエリ "tool\editor" に対して
-        // Substring マッチ: "editor".find("tool\editor") → None（クエリが長い）
-        // 両方ともパスマッチのみ → path_score で順序が決まる
-        // path_score は byte_position で比較 — 同じパスプレフィックスなのでスコア同等
-        // タイブレーク: lower_name 昇順 → "app" < "editor"
-        assert_eq!(results2[0].name, "app");
-        assert_eq!(results2[1].name, "editor");
-    }
-
-    #[test]
-    fn path_match_slash_normalized() {
-        let entries = vec![make_entry("app", "C:\\tool\\editor\\app.exe")];
-        let mut engine = SearchEngine::new(entries);
-        // `/` で入力しても `\` に正規化されてマッチする
-        let results = engine.search("tool/editor", 8, &empty_history(), SearchMode::Substring);
-        assert_eq!(results.len(), 1);
-        assert_eq!(results[0].name, "app");
-    }
-
-    #[test]
-    fn path_match_receives_history_boost() {
-        let entries = vec![
-            make_entry("app1", "C:\\tool\\editor\\app1.exe"),
-            make_entry("app2", "C:\\tool\\editor\\app2.exe"),
-        ];
-        let mut history = HistoryStore::load();
-        for _ in 0..5 {
-            history.record_launch("C:\\tool\\editor\\app1.exe", "");
-        }
-        let mut engine = SearchEngine::new(entries);
-        let results = engine.search("tool\\editor", 8, &history, SearchMode::Substring);
-        assert_eq!(results.len(), 2);
-        // app1 は history boost で上位
-        assert_eq!(results[0].name, "app1");
-        assert_eq!(results[1].name, "app2");
-    }
-
-    #[test]
-    fn path_match_incremental_cache_monotonic() {
-        let entries = vec![
-            make_entry("app", "C:\\tool\\editor\\app.exe"),
-            make_entry("other", "C:\\other\\other.exe"),
-        ];
-        let mut engine = SearchEngine::new(entries);
-        let h = empty_history();
-
-        // 1回目: "tool\\" でパスマッチ
-        let r1 = engine.search("tool\\", 8, &h, SearchMode::Substring);
-        assert_eq!(r1.len(), 1);
-        assert_eq!(r1[0].name, "app");
-
-        // 2回目: "tool\\ed" に拡張 → パス区切りを含むため incremental は無効化される
-        //（decide_incremental の !has_path_sep ガード）。fresh scan と一致することを検証。
-        let r2 = engine.search("tool\\ed", 8, &h, SearchMode::Substring);
-        assert_eq!(r2.len(), 1);
-        assert_eq!(r2[0].name, "app");
-
-        // 比較: fresh engine と同じ結果になること
-        let mut fresh = SearchEngine::new(vec![
-            make_entry("app", "C:\\tool\\editor\\app.exe"),
-            make_entry("other", "C:\\other\\other.exe"),
-        ]);
-        let fresh_result = fresh.search("tool\\ed", 8, &h, SearchMode::Substring);
-        assert_eq!(r2.len(), fresh_result.len());
-        assert_eq!(r2[0].name, fresh_result[0].name);
-    }
-
-    #[test]
-    fn path_match_no_match_returns_empty() {
-        let entries = vec![make_entry("app", "C:\\tool\\editor\\app.exe")];
-        let mut engine = SearchEngine::new(entries);
-        let results = engine.search("xyz\\abc", 8, &empty_history(), SearchMode::Substring);
-        assert!(results.is_empty());
-    }
-
-    #[test]
-    fn path_match_fuzzy_mode_skips_bitmask_prefilter() {
-        // name="zzz" はクエリ "tool\\editor" の文字 (t,o,l,e,d,i,r) を含まない
-        // → 通常ならビットマスクで除外されるが、has_path_sep でスキップされパスマッチする
-        let entries = vec![make_entry("zzz", "C:\\tool\\editor\\zzz.exe")];
-        let mut engine = SearchEngine::new(entries);
-        let results = engine.search("tool\\editor", 8, &empty_history(), SearchMode::Fuzzy);
-        assert_eq!(results.len(), 1);
-        assert_eq!(results[0].name, "zzz");
-    }
-
-    #[test]
-    fn path_match_yen_sign_normalized() {
-        // ¥（U+00A5）は日本語 Windows でバックスラッシュとして使われる
-        let entries = vec![make_entry("app", "C:\\tool\\editor\\app.exe")];
-        let mut engine = SearchEngine::new(entries);
-        let results = engine.search("tool\u{00a5}editor", 8, &empty_history(), SearchMode::Substring);
-        assert_eq!(results.len(), 1);
-        assert_eq!(results[0].name, "app");
-    }
-
-    #[test]
-    fn path_match_consecutive_spaces_preserved() {
-        // パス成分に連続スペースを含む場合、normalize_query() で潰されない
-        let entries = vec![make_entry("app", "C:\\My  Tools\\app.exe")];
-        let mut engine = SearchEngine::new(entries);
-        let results = engine.search("My  Tools\\", 8, &empty_history(), SearchMode::Substring);
-        assert_eq!(results.len(), 1);
-        assert_eq!(results[0].name, "app");
-    }
-
-    #[test]
-    fn path_match_incremental_disabled_avoids_accent_false_negative() {
-        // incremental cache が有効だった場合の false negative を検証する。
-        // entry path に "café" を含み、"cafe\\" (no accent) → "café\\" (with accent)
-        // の遷移で norm_query は両方 "cafe\\" だが、path_query は異なる。
-        // incremental 無効化により、full scan で正しくマッチする。
-        let entries = vec![
-            make_entry("app", "C:\\café\\app.exe"),
-            make_entry("other", "C:\\other\\other.exe"),
-        ];
-        let mut engine = SearchEngine::new(entries);
-        let h = empty_history();
-
-        // 1回目: "cafe\\" — normalized_key は "c:\café\app.exe" (accent preserved)
-        // path_query "cafe\\" は "café" にマッチしない
-        let r1 = engine.search("cafe\\", 8, &h, SearchMode::Substring);
-        assert!(r1.is_empty());
-
-        // 2回目: "café\\" — path_query "café\\" は "café" にマッチすべき
-        // incremental が有効だと前回の空結果を再利用して false negative になる
-        let r2 = engine.search("café\\", 8, &h, SearchMode::Substring);
-        assert_eq!(r2.len(), 1);
-        assert_eq!(r2[0].name, "app");
-    }
-
-    #[test]
-    fn path_match_history_key_unified_across_separators() {
-        // tool/editor と tool\editor で履歴バケットが統一される
-        let entries = vec![
-            make_entry("app1", "C:\\tool\\editor\\app1.exe"),
-            make_entry("app2", "C:\\tool\\editor\\app2.exe"),
-        ];
-        let mut history = HistoryStore::load();
-        // tool/editor（スラッシュ）で起動記録
-        history.record_launch("C:\\tool\\editor\\app1.exe", "tool/editor");
-
-        let mut engine = SearchEngine::new(entries);
-        // tool\editor（バックスラッシュ）で検索 → 履歴が効くべき
-        let results = engine.search("tool\\editor", 8, &history, SearchMode::Substring);
-        assert_eq!(results.len(), 2);
-        assert_eq!(results[0].name, "app1"); // history boost で上位
-    }
-}
+mod tests;
